refactor(api): Share pagination module
